app/requirements.txt
fastapi
uvicorn
psycopg2-binary
# psycopg2
redis
uvicorn[standard]
